+++ src/UglyToad.PdfPig/Writer/Xmp/XmpWriter.cs
@@ -38,7 +38,7 @@ namespace UglyToad.PdfPig.Writer.Xmp
         private const string PdfAIdentificationExtensionPrefix = "pdfaid";
         private const string PdfAIdentificationExtensionNamespace = "http://www.aiim.org/pdfa/ns/id/";
         
-        public static StreamToken GenerateXmpStream(PdfDocumentBuilder.DocumentInformationBuilder builder, decimal version,
+        public static StreamToken GenerateXmpStream(PdfDocumentBuilder.DocumentInformationBuilder builder, double version,
             PdfAStandard standard, XDocument additionalXmpMetadata)
         {
             XNamespace xmpMeta = XmpMetaNamespace;
